requirements.txt
pandas
requests
tqdm
PyPDF2
pdfminer.six
boxsdk